# internal/http/openapi_spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GoClaw Gateway API",
    "description": "PostgreSQL multi-tenant AI agent gateway with WebSocket RPC + HTTP API.\n\n## Authentication\n\nAll endpoints require a Bearer token in the `Authorization` header:\n\n```\nAuthorization: Bearer <gateway-token-or-api-key>\n```\n\nYou can use either the **gateway token** (grants admin access) or an **API key** created via the API Keys endpoints (grants scoped access).\n\nIf no token is configured on the server, authentication is disabled (backward compatibility).\n\n## Common Headers\n\n| Header | Description |\n|--------|-------------|\n| `X-GoClaw-User-Id` | External user ID for multi-tenant context |\n| `X-GoClaw-Agent-Id` | Target agent ID (alternative to model prefix) |\n| `X-GoClaw-Tenant-Id` | Tenant scope — UUID or slug (gateway token / cross-tenant API keys) |\n| `Accept-Language` | Locale for error messages (`en`, `vi`, `zh`) |\n\n## WebSocket Protocol\n\nConnect via `POST /ws` (upgrade). Protocol v3 uses frame types: `req`, `res`, `event`.\nFirst request must be `connect` with `{\"token\": \"...\", \"user_id\": \"...\", \"locale\": \"en\"}`.",
    "version": "0.2.0",
    "contact": {
      "name": "GoClaw",
      "url": "https://github.com/nextlevelbuilder/goclaw"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/providers": {
      "get": {
        "tags": [
          "Providers"
        ],
        "summary": "List LLM providers",
        "responses": {
          "200": {
            "description": "Provider list (API keys masked)"
          }
        }
      },
      "post": {
        "tags": [
          "Providers"
        ],
        "summary": "Create LLM provider",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Provider created"
          }
        }
      }
    },
    "/v1/providers/{id}": {
      "get": {
        "tags": [
          "Providers"
        ],
        "summary": "Get provider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider details (API key masked)"
          }
        }
      },
      "put": {
        "tags": [
          "Providers"
        ],
        "summary": "Update provider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Providers"
        ],
        "summary": "Delete provider",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider deleted"
          }
        }
      }
    },
    "/v1/providers/{id}/models": {
      "get": {
        "tags": [
          "Providers"
        ],
        "summary": "List available models",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model list with optional reasoning capability metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "models": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "reasoning": {
                            "type": "object",
                            "properties": {
                              "levels": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              },
                              "default_effort": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "required": [
                          "id"
                        ]
                      }
                    },
                    "reasoning_defaults": {
                      "type": "object",
                      "description": "Provider-owned default reasoning policy returned when the provider has saved defaults and at least one discovered model exposes reasoning capability metadata.",
                      "properties": {
                        "effort": {
                          "type": "string",
                          "enum": [
                            "off",
                            "auto",
                            "none",
                            "minimal",
                            "low",
                            "medium",
                            "high",
                            "xhigh"
                          ]
                        },
                        "fallback": {
                          "type": "string",
                          "enum": [
                            "downgrade",
                            "off",
                            "provider_default"
                          ]
                        }
                      }
                    }
                  },
                  "required": [
                    "models"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/providers/{id}/verify": {
      "post": {
        "tags": [
          "Providers"
        ],
        "summary": "Verify provider connection",
        "description": "Empty body or omitted model = ping mode (provider registered/reachable check). With model = full chat-verify against the upstream provider.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string",
                    "description": "Optional. Omit (or send empty body) for ping mode."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Gateway token or API key"
      }
    },
    "schemas": {
...
      "ProviderInput": {
        "type": "object",
        "required": [
          "name",
          "provider_type"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Unique provider slug"
          },
          "display_name": {
            "type": "string"
          },
          "provider_type": {
            "type": "string",
            "enum": [
              "anthropic_native",
              "openai_compat",
              "gemini_native",
              "openrouter",
              "groq",
              "deepseek",
              "mistral",
              "xai",
              "minimax_native",
              "cohere",
              "perplexity",
              "dashscope",
              "bailian",
              "chatgpt_oauth",
              "claude_cli",
              "yescale",
              "zai",
              "zai_coding",
              "vertex",
              "ollama",
              "ollama_cloud",
              "acp",
              "novita",
              "byteplus",
              "byteplus_coding"
            ]
          },
          "api_base": {
            "type": "string"
          },
          "api_key": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean",
            "default": true
          },
          "settings": {
            "type": "object",
            "description": "Provider-specific settings such as provider-owned Codex pool defaults and reasoning defaults.",
            "properties": {
              "reasoning_defaults": {
                "type": "object",
                "properties": {
                  "effort": {
                    "type": "string",
                    "enum": [
                      "off",
                      "auto",
                      "none",
                      "minimal",
                      "low",
                      "medium",
                      "high",
                      "xhigh"
                    ]
                  },
                  "fallback": {
                    "type": "string",
                    "enum": [
                      "downgrade",
                      "off",
                      "provider_default"
                    ]
                  }
                }
              },
              "codex_pool": {
                "type": "object",
                "properties": {
                  "strategy": {
                    "type": "string",
                    "enum": [
                      "round_robin",
                      "priority_order"
                    ]
                  },
                  "extra_provider_names": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
...
  }
}